requirements.txt
discord.py[voice]
discord-eprompt
python-dateutil
fuzzywuzzy
wallet-sdk
requests
voluptuous
pyjwt
twine
